prep-for-a-meeting/requirements.txt
langchain-exa==0.0.1
exa_py=1.0.7
python-dotenv=1.0.1
